refactor!: Refactor out all fixed-width types properties (#39)
We were erroneously using fixed-width types for property
storage and access in many places, which had the potential
to cause unintended stack smashing. This commit refactors
all integer property types and usage to basic g(u)int types.

Refs: BARTON-143

+++ api/c/src/device-service-device.c
@@ -36,7 +36,7 @@ struct _BDeviceServiceDevice
 
     gchar *uuid;
     gchar *deviceClass;
-    uint8_t deviceClassVersion;
+    guint deviceClassVersion;
     gchar *uri;
     gchar *managingDeviceDriver;
     GList *endpoints;
